Organize components into folders and format styles
