+++ src/App.css
@@ -1,16 +1,103 @@
-.app{min-height:100vh;background:var(--bg);color:var(--text);}
-.container{max-width:1200px;margin:0 auto;padding:24px;display:flex;flex-direction:column;gap:24px;}
-.banner{padding:12px;border-radius:10px;border:1px solid var(--accent);background:var(--accent-subtle);color:var(--text);}
-.grid-wrap{display:grid;grid-template-columns:repeat(3,minmax(0,1fr));gap:24px;}
-@media(max-width:1100px){.grid-wrap{grid-template-columns:repeat(2,minmax(0,1fr));}}
-@media(max-width:760px){.grid-wrap{grid-template-columns:1fr;}}
-.palette-info{opacity:0.7;}
-.palette-info span{color:var(--accent);}
-.workspace-bar{display:grid;grid-template-columns:1fr auto auto;gap:8px;align-items:center;}
-.workspace-create{display:grid;grid-template-columns:1fr auto;gap:8px;margin-top:10px;}
-.workspace-bar-select,.workspace-bar-input{background:transparent;color:var(--text);border:1px solid var(--accent);border-radius:8px;padding:6px 8px;}
-.muted{opacity:0.6;}
-.accent{color:var(--accent);}
-.palette-royalViolet{--bg:#1B1B2F;--card:#2C2C3E;--text:#F1F1F8;--accent:#9D4EDD;--accent-subtle:#C77DFF;}
-.palette-coralSpark{--bg:#202124;--card:#2F3136;--text:#EEEEEE;--accent:#FF6B6B;--accent-subtle:#FF9F80;}
-.palette-neonCyan{--bg:#0D1B2A;--card:#1B263B;--text:#E0E1DD;--accent:#00B4D8;--accent-subtle:#90E0EF;}
+.app {
+  min-height: 100vh;
+  background: var(--bg);
+  color: var(--text);
+}
+
+.container {
+  max-width: 1200px;
+  margin: 0 auto;
+  padding: 24px;
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+}
+
+.banner {
+  padding: 12px;
+  border-radius: 10px;
+  border: 1px solid var(--accent);
+  background: var(--accent-subtle);
+  color: var(--text);
+}
+
+.grid-wrap {
+  display: grid;
+  grid-template-columns: repeat(3, minmax(0, 1fr));
+  gap: 24px;
+}
+
+@media (max-width: 1100px) {
+  .grid-wrap {
+    grid-template-columns: repeat(2, minmax(0, 1fr));
+  }
+}
+
+@media (max-width: 760px) {
+  .grid-wrap {
+    grid-template-columns: 1fr;
+  }
+}
+
+.palette-info {
+  opacity: 0.7;
+}
+
+.palette-info span {
+  color: var(--accent);
+}
+
+.workspace-bar {
+  display: grid;
+  grid-template-columns: 1fr auto auto;
+  gap: 8px;
+  align-items: center;
+}
+
+.workspace-create {
+  display: grid;
+  grid-template-columns: 1fr auto;
+  gap: 8px;
+  margin-top: 10px;
+}
+
+.workspace-bar-select,
+.workspace-bar-input {
+  background: transparent;
+  color: var(--text);
+  border: 1px solid var(--accent);
+  border-radius: 8px;
+  padding: 6px 8px;
+}
+
+.muted {
+  opacity: 0.6;
+}
+
+.accent {
+  color: var(--accent);
+}
+
+.palette-royalViolet {
+  --bg: #1B1B2F;
+  --card: #2C2C3E;
+  --text: #F1F1F8;
+  --accent: #9D4EDD;
+  --accent-subtle: #C77DFF;
+}
+
+.palette-coralSpark {
+  --bg: #202124;
+  --card: #2F3136;
+  --text: #EEEEEE;
+  --accent: #FF6B6B;
+  --accent-subtle: #FF9F80;
+}
+
+.palette-neonCyan {
+  --bg: #0D1B2A;
+  --card: #1B263B;
+  --text: #E0E1DD;
+  --accent: #00B4D8;
+  --accent-subtle: #90E0EF;
+}
