+++ src/components/goals-overview/goals-overview.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import List from "./List";
+import List from "../list/list";
 
 export default function GoalsOverview(props) {
   return <List title="Goals" type="goals" {...props} />;
