+++ src/components/header/header.css
@@ -1,0 +1,54 @@
+.header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.header-title {
+  margin: 0;
+  color: var(--text);
+}
+
+.header-controls {
+  display: flex;
+  gap: 12px;
+  align-items: center;
+}
+
+.theme-label {
+  opacity: 0.8;
+}
+
+.theme-select {
+  background: transparent;
+  color: var(--text);
+  border: 1px solid var(--accent);
+  border-radius: 8px;
+  padding: 4px 6px;
+}
+
+.theme-select option {
+  color: #000;
+}
+
+.auth-panel {
+  display: flex;
+  gap: 12px;
+  align-items: center;
+}
+
+.auth-input {
+  background: transparent;
+  color: var(--text);
+  border: 1px solid var(--accent);
+  border-radius: 8px;
+  padding: 6px 8px;
+}
+
+.greeting {
+  opacity: 0.85;
+}
+
+.auth-error {
+  color: var(--accent);
+}
