+++ src/components/people-overview/people-overview.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import List from "./List";
+import List from "../list/list";
 
 export default function PeopleOverview(props) {
   return <List title="Tracked People Projects" type="people" {...props} />;
