+++ src/components/progress-overview/progress-overview.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import List from "./List";
+import List from "../list/list";
 
 export default function ProgressOverview(props) {
   return <List title="Track Progress" type="progress" {...props} />;
