--- src/components/ui.css
@@ -1,9 +1,0 @@
-.card{background:var(--card);border:1px solid var(--accent);border-radius:16px;padding:16px;box-shadow:0 6px 20px rgba(0,0,0,0.25);}
-.card-header{display:flex;justify-content:space-between;align-items:center;margin-bottom:12px;}
-.card-title{margin:0;color:var(--text);}
-.btn{padding:8px 12px;border-radius:10px;border:1px solid var(--accent);background:var(--accent-subtle);color:var(--text);cursor:pointer;}
-.btn-subtle{background:transparent;}
-.btn:disabled{cursor:not-allowed;opacity:0.6;}
-.add-row{display:grid;grid-template-columns:1fr auto;gap:8px;margin-top:12px;}
-.add-row input{background:transparent;color:inherit;border:1px solid rgba(255,255,255,.2);border-radius:8px;padding:8px 10px;}
-.add-row button{border-radius:8px;border:1px solid rgba(255,255,255,.3);padding:8px 10px;}
